Update TileDB dependencies (#2034)
* Update Zstd to latest stable version

* Update openSSL to latest stable version

* Update LZ4 to latest stable version

* Update Curl to latest stable version

* Update Clipp to latest stable version

* Fix Curl compilation on Windows

* Update Catch to latest stable version

* Fix warnings caused by Catch update

+++ test/src/unit-threadpool.cc
@@ -174,7 +174,7 @@ TEST_CASE(
     }
 
     REQUIRE(result == num_ok);
-    REQUIRE(num_cancelled == (tasks.size() - num_ok));
+    REQUIRE(num_cancelled == ((int64_t)tasks.size() - num_ok));
   }
 }
